+++ core/src/muxing/boxed.rs
@@ -1,27 +1,35 @@
 use crate::muxing::StreamMuxerEvent;
 use crate::StreamMuxer;
 use fnv::FnvHashMap;
+use futures::{AsyncRead, AsyncWrite};
 use parking_lot::Mutex;
+use std::fmt;
 use std::io;
+use std::io::{IoSlice, IoSliceMut};
+use std::pin::Pin;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::task::{Context, Poll};
 
 /// Abstract `StreamMuxer`.
 pub struct StreamMuxerBox {
     inner: Box<
-        dyn StreamMuxer<Substream = usize, OutboundSubstream = usize, Error = io::Error>
+        dyn StreamMuxer<Substream = SubstreamBox, OutboundSubstream = usize, Error = io::Error>
             + Send
             + Sync,
     >,
 }
+
+/// Abstract type for asynchronous reading and writing.
+///
+/// A [`SubstreamBox`] erases the concrete type it is given and only retains its `AsyncRead`
+/// and `AsyncWrite` capabilities.
+pub struct SubstreamBox(Box<dyn AsyncReadWrite + Send + Unpin>);
 
 struct Wrap<T>
 where
     T: StreamMuxer,
 {
     inner: T,
-    substreams: Mutex<FnvHashMap<usize, T::Substream>>,
-    next_substream: AtomicUsize,
     outbound: Mutex<FnvHashMap<usize, T::OutboundSubstream>>,
     next_outbound: AtomicUsize,
 }
@@ -29,8 +37,9 @@ where
 impl<T> StreamMuxer for Wrap<T>
 where
     T: StreamMuxer,
+    T::Substream: Send + Unpin + 'static,
 {
-    type Substream = usize; // TODO: use a newtype
+    type Substream = SubstreamBox;
     type OutboundSubstream = usize; // TODO: use a newtype
     type Error = io::Error;
 
@@ -48,9 +57,9 @@ where
             Poll::Ready(Err(err)) => return Poll::Ready(Err(err.into())),
         };
 
-        let id = self.next_substream.fetch_add(1, Ordering::Relaxed);
-        self.substreams.lock().insert(id, substream);
-        Poll::Ready(Ok(StreamMuxerEvent::InboundSubstream(id)))
+        Poll::Ready(Ok(StreamMuxerEvent::InboundSubstream(SubstreamBox::new(
+            substream,
+        ))))
     }
 
     #[inline]
@@ -76,9 +85,8 @@ where
             Poll::Ready(Ok(s)) => s,
             Poll::Ready(Err(err)) => return Poll::Ready(Err(err.into())),
         };
-        let id = self.next_substream.fetch_add(1, Ordering::Relaxed);
-        self.substreams.lock().insert(id, substream);
-        Poll::Ready(Ok(id))
+
+        Poll::Ready(Ok(SubstreamBox::new(substream)))
     }
 
     #[inline]
@@ -86,63 +94,6 @@ where
         let mut list = self.outbound.lock();
         self.inner
             .destroy_outbound(list.remove(&substream).unwrap())
-    }
-
-    #[inline]
-    fn read_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-        buf: &mut [u8],
-    ) -> Poll<Result<usize, Self::Error>> {
-        let mut list = self.substreams.lock();
-        self.inner
-            .read_substream(cx, list.get_mut(s).unwrap(), buf)
-            .map_err(|e| e.into())
-    }
-
-    #[inline]
-    fn write_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-        buf: &[u8],
-    ) -> Poll<Result<usize, Self::Error>> {
-        let mut list = self.substreams.lock();
-        self.inner
-            .write_substream(cx, list.get_mut(s).unwrap(), buf)
-            .map_err(|e| e.into())
-    }
-
-    #[inline]
-    fn flush_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-    ) -> Poll<Result<(), Self::Error>> {
-        let mut list = self.substreams.lock();
-        self.inner
-            .flush_substream(cx, list.get_mut(s).unwrap())
-            .map_err(|e| e.into())
-    }
-
-    #[inline]
-    fn shutdown_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-    ) -> Poll<Result<(), Self::Error>> {
-        let mut list = self.substreams.lock();
-        self.inner
-            .shutdown_substream(cx, list.get_mut(s).unwrap())
-            .map_err(|e| e.into())
-    }
-
-    #[inline]
-    fn destroy_substream(&self, substream: Self::Substream) {
-        let mut list = self.substreams.lock();
-        self.inner
-            .destroy_substream(list.remove(&substream).unwrap())
     }
 
     #[inline]
@@ -157,12 +108,10 @@ impl StreamMuxerBox {
     where
         T: StreamMuxer + Send + Sync + 'static,
         T::OutboundSubstream: Send,
-        T::Substream: Send,
+        T::Substream: Send + Unpin + 'static,
     {
         let wrap = Wrap {
             inner: muxer,
-            substreams: Mutex::new(Default::default()),
-            next_substream: AtomicUsize::new(0),
             outbound: Mutex::new(Default::default()),
             next_outbound: AtomicUsize::new(0),
         };
@@ -174,7 +123,7 @@ impl StreamMuxerBox {
 }
 
 impl StreamMuxer for StreamMuxerBox {
-    type Substream = usize; // TODO: use a newtype
+    type Substream = SubstreamBox;
     type OutboundSubstream = usize; // TODO: use a newtype
     type Error = io::Error;
 
@@ -206,50 +155,81 @@ impl StreamMuxer for StreamMuxerBox {
     }
 
     #[inline]
-    fn read_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-        buf: &mut [u8],
-    ) -> Poll<Result<usize, Self::Error>> {
-        self.inner.read_substream(cx, s, buf)
-    }
-
-    #[inline]
-    fn write_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-        buf: &[u8],
-    ) -> Poll<Result<usize, Self::Error>> {
-        self.inner.write_substream(cx, s, buf)
-    }
-
-    #[inline]
-    fn flush_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-    ) -> Poll<Result<(), Self::Error>> {
-        self.inner.flush_substream(cx, s)
-    }
-
-    #[inline]
-    fn shutdown_substream(
-        &self,
-        cx: &mut Context<'_>,
-        s: &mut Self::Substream,
-    ) -> Poll<Result<(), Self::Error>> {
-        self.inner.shutdown_substream(cx, s)
-    }
-
-    #[inline]
-    fn destroy_substream(&self, s: Self::Substream) {
-        self.inner.destroy_substream(s)
-    }
-
-    #[inline]
     fn poll_close(&self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
         self.inner.poll_close(cx)
+    }
+}
+
+impl SubstreamBox {
+    /// Construct a new [`SubstreamBox`] from something that implements [`AsyncRead`] and [`AsyncWrite`].
+    pub fn new<S: AsyncRead + AsyncWrite + Send + Unpin + 'static>(stream: S) -> Self {
+        Self(Box::new(stream))
+    }
+}
+
+impl fmt::Debug for SubstreamBox {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "SubstreamBox({})", self.0.type_name())
+    }
+}
+
+/// Workaround because Rust does not allow `Box<dyn AsyncRead + AsyncWrite>`.
+trait AsyncReadWrite: AsyncRead + AsyncWrite + Unpin {
+    /// Helper function to capture the erased inner type.
+    ///
+    /// Used to make the [`Debug`] implementation of [`SubstreamBox`] more useful.
+    fn type_name(&self) -> &'static str;
+}
+
+impl<S> AsyncReadWrite for S
+where
+    S: AsyncRead + AsyncWrite + Unpin,
+{
+    fn type_name(&self) -> &'static str {
+        std::any::type_name::<S>()
+    }
+}
+
+impl AsyncRead for SubstreamBox {
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &mut [u8],
+    ) -> Poll<std::io::Result<usize>> {
+        Pin::new(&mut self.get_mut().0).poll_read(cx, buf)
+    }
+
+    fn poll_read_vectored(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        bufs: &mut [IoSliceMut<'_>],
+    ) -> Poll<std::io::Result<usize>> {
+        Pin::new(&mut self.get_mut().0).poll_read_vectored(cx, bufs)
+    }
+}
+
+impl AsyncWrite for SubstreamBox {
+    fn poll_write(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &[u8],
+    ) -> Poll<std::io::Result<usize>> {
+        Pin::new(&mut self.get_mut().0).poll_write(cx, buf)
+    }
+
+    fn poll_write_vectored(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        bufs: &[IoSlice<'_>],
+    ) -> Poll<std::io::Result<usize>> {
+        Pin::new(&mut self.get_mut().0).poll_write_vectored(cx, bufs)
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
+        Pin::new(&mut self.get_mut().0).poll_flush(cx)
+    }
+
+    fn poll_close(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
+        Pin::new(&mut self.get_mut().0).poll_close(cx)
     }
 }