+++ protocols/gossipsub/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.39.0 [unreleased]
+
+- Update to `libp2p-core` `v0.34.0`.
+
+- Update to `libp2p-swarm` `v0.37.0`.
+
 # 0.38.1
 
 - Fix duplicate connection id. See [PR 2702].